# statistics_manager_service/openapi/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: vasco.m.campos@inesctec.pt
    name: Vasco Campos
  description: "This is the Statistics Manager Service OpenAPI definition. The Statistics\
    \ Manager Service processes energy data in order to provide the user with useful\
    \ statistics (e.g. energy cost for the day, injection for the month, consumption\
    \ for the year, CO2 footprint). The interactions here represented are based on\
    \ this [Canvas](https://gitlab.inesctec.pt/cpes/european-projects/interconnect/hems/hems-documentation/-/blob/master/Microservices/Statistics-Manager-Service.adoc)"
  title: Statistics Manager Service
  version: 0.1.1
servers:
- url: https://interconnect-dev.inesctec.pt/api/statistics
security:
- {}
- Bearer: []
...
paths:
...
  /power-consumption:
    description: Retrieve the power consumption for several users.
    get:
      operationId: power_consumption_get
      parameters:
      - explode: false
        in: header
        name: X-Correlation-ID
        required: true
        schema:
          $ref: '#/components/schemas/CorrelationId'
        style: simple
      - explode: false
        in: header
        name: Authorization
        required: false
        schema:
          $ref: '#/components/schemas/Authorization'
        style: simple
      - explode: true
        in: query
        name: user_ids
        required: true
        schema:
          example:
          - userId1
          - userId2
          items:
            $ref: '#/components/schemas/UserId'
          minItems: 1
          type: array
        style: form
      - explode: true
        in: query
        name: group_by
        required: false
        schema:
          default: daily
          enum:
          - 15_mins
          - hourly
          - daily
          - weekly
          - monthly
          example: daily
          type: string
        style: form
      - explode: true
        in: query
        name: start_date
        required: true
        schema:
          $ref: '#/components/schemas/Timestamp'
        style: form
      - explode: true
        in: query
        name: end_date
        required: false
        schema:
          $ref: '#/components/schemas/Timestamp'
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/PowerMetrics'
                type: array
          description: Success response The timestamp field is the first timestamp
            of the returned values.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "400":
          content:
            application/json:
              examples:
                Timestamp error:
                  value:
                    error: Reason for bad request error
                Missing X-Correlation-ID header:
                  value:
                    error: missing X-Correlation-ID header
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "401":
          content:
            application/json:
              examples:
                Unauthenticated:
                  value:
                    error: error in authorization token
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthenticated or invalid credentials
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "403":
          content:
            application/json:
              examples:
                Forbidden:
                  value:
                    error: you are not authorized to access this information
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden to access the information requested
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "404":
          content:
            application/json:
              examples:
                User Id not found:
                  value:
                    error: user with id "userId" not found
              schema:
                $ref: '#/components/schemas/Error'
          description: User id not found
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
      summary: Retrieve the power consumption for several users. The "group_by" variable
        is used to compute the average value between that interval
      tags:
      - Power Queries
      x-openapi-router-controller: statistics_manager_service.controllers.power_queries_controller
  /power-consumption-last:
    description: Retrieve the last power value of the consumption for several users.
    get:
      operationId: power_consumption_last_get
      parameters:
      - explode: false
        in: header
        name: X-Correlation-ID
        required: true
        schema:
          $ref: '#/components/schemas/CorrelationId'
        style: simple
      - explode: false
        in: header
        name: Authorization
        required: false
        schema:
          $ref: '#/components/schemas/Authorization'
        style: simple
      - explode: true
        in: query
        name: user_ids
        required: true
        schema:
          example:
          - userId1
          - userId2
          items:
            $ref: '#/components/schemas/UserId'
          minItems: 1
          type: array
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/InstantPowerMetric'
                type: array
          description: Success response
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "400":
          content:
            application/json:
              examples:
                Timestamp error:
                  value:
                    error: Reason for bad request error
                Missing X-Correlation-ID header:
                  value:
                    error: missing X-Correlation-ID header
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "401":
          content:
            application/json:
              examples:
                Unauthenticated:
                  value:
                    error: error in authorization token
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthenticated or invalid credentials
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "403":
          content:
            application/json:
              examples:
                Forbidden:
                  value:
                    error: you are not authorized to access this information
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden to access the information requested
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "404":
          content:
            application/json:
              examples:
                User Id not found:
                  value:
                    error: user with id "userId" not found
              schema:
                $ref: '#/components/schemas/Error'
          description: User id not found
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
      summary: Retrieve the last power value of the consumption for several users.
      tags:
      - Power Queries
      x-openapi-router-controller: statistics_manager_service.controllers.power_queries_controller
  /power-injection:
    description: Retrieve the power injection for several users.
    get:
      operationId: power_injection_get
      parameters:
      - explode: false
        in: header
        name: X-Correlation-ID
        required: true
        schema:
          $ref: '#/components/schemas/CorrelationId'
        style: simple
      - explode: false
        in: header
        name: Authorization
        required: false
        schema:
          $ref: '#/components/schemas/Authorization'
        style: simple
      - explode: true
        in: query
        name: user_ids
        required: true
        schema:
          example:
          - userId1
          - userId2
          items:
            $ref: '#/components/schemas/UserId'
          minItems: 1
          type: array
        style: form
      - explode: true
        in: query
        name: group_by
        required: false
        schema:
          default: daily
          enum:
          - 15_mins
          - hourly
          - daily
          - weekly
          - monthly
          example: daily
          type: string
        style: form
      - explode: true
        in: query
        name: start_date
        required: true
        schema:
          $ref: '#/components/schemas/Timestamp'
        style: form
      - explode: true
        in: query
        name: end_date
        required: false
        schema:
          $ref: '#/components/schemas/Timestamp'
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/PowerMetrics'
                type: array
          description: Success response The timestamp field is the first timestamp
            of the returned values.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "400":
          content:
            application/json:
              examples:
                Timestamp error:
                  value:
                    error: Reason for bad request error
                Missing X-Correlation-ID header:
                  value:
                    error: missing X-Correlation-ID header
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "401":
          content:
            application/json:
              examples:
                Unauthenticated:
                  value:
                    error: error in authorization token
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthenticated or invalid credentials
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "403":
          content:
            application/json:
              examples:
                Forbidden:
                  value:
                    error: you are not authorized to access this information
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden to access the information requested
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "404":
          content:
            application/json:
              examples:
                User Id not found:
                  value:
                    error: user with id "userId" not found
              schema:
                $ref: '#/components/schemas/Error'
          description: User id not found
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
      summary: Retrieve the power injection for several users. The "group_by" variable
        is used to compute the average value between that interval
      tags:
      - Power Queries
      x-openapi-router-controller: statistics_manager_service.controllers.power_queries_controller
  /power-injection-last:
    description: Retrieve the last power value of the injection for several users.
    get:
      operationId: power_injection_last_get
      parameters:
      - explode: false
        in: header
        name: X-Correlation-ID
        required: true
        schema:
          $ref: '#/components/schemas/CorrelationId'
        style: simple
      - explode: false
        in: header
        name: Authorization
        required: false
        schema:
          $ref: '#/components/schemas/Authorization'
        style: simple
      - explode: true
        in: query
        name: user_ids
        required: true
        schema:
          example:
          - userId1
          - userId2
          items:
            $ref: '#/components/schemas/UserId'
          minItems: 1
          type: array
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/InstantPowerMetric'
                type: array
          description: Success response
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "400":
          content:
            application/json:
              examples:
                Timestamp error:
                  value:
                    error: Reason for bad request error
                Missing X-Correlation-ID header:
                  value:
                    error: missing X-Correlation-ID header
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "401":
          content:
            application/json:
              examples:
                Unauthenticated:
                  value:
                    error: error in authorization token
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthenticated or invalid credentials
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "403":
          content:
            application/json:
              examples:
                Forbidden:
                  value:
                    error: you are not authorized to access this information
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden to access the information requested
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "404":
          content:
            application/json:
              examples:
                User Id not found:
                  value:
                    error: user with id "userId" not found
              schema:
                $ref: '#/components/schemas/Error'
          description: User id not found
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
      summary: Retrieve the last power value of the injection for several users.
      tags:
      - Power Queries
      x-openapi-router-controller: statistics_manager_service.controllers.power_queries_controller
...
components:
...
  headers:
    CorrelationId:
      explode: false
      required: true
      schema:
        $ref: '#/components/schemas/CorrelationId'
      style: simple
...
  schemas:
    CorrelationId:
      format: uuid
      type: string
    UserId:
      example: userId1
      maxLength: 10
      minLength: 10
      title: user_id
      type: string
    Authorization:
      format: byte
      type: string
...
    Timestamp:
      example: 2023-01-20T00:00:00Z
      format: date-time
      title: timestamp
      type: string
...
    PowerUnits:
      enum:
      - W
      - kW
      example: kW
      title: PowerUnits
      type: string
...
    PowerMetric:
      description: \"timestamp\" and \"slotNumber\" field are incremented by the query
        parameter \"timeInterval\" value = grid_consumed_value + self_consumed_value
        value = thermal_value + other_value
      example:
        units: kW
        value: 12.5
        timestamp: 2023-01-20T00:00:00Z
        slot_number: 2
      properties:
        timestamp:
          example: 2023-01-20T00:00:00Z
          format: date-time
          title: timestamp
          type: string
        value:
          example: 12.5
          format: float
          minimum: 0
          title: value
          type: number
        units:
          $ref: '#/components/schemas/PowerUnits'
        slot_number:
          example: 2
          format: int64
          minimum: 1
          title: slot_number
          type: number
      required:
      - slot_number
      - timestamp
      - units
      - value
      title: PowerMetric
      type: object
    PowerMetrics:
      example:
        user_id: userId1
        metrics:
        - units: kW
          value: 12.5
          timestamp: 2023-01-20T00:00:00Z
          slot_number: 2
        - units: kW
          value: 12.5
          timestamp: 2023-01-20T00:00:00Z
          slot_number: 2
      properties:
        user_id:
          example: userId1
          maxLength: 10
          minLength: 10
          title: user_id
          type: string
        metrics:
          items:
            $ref: '#/components/schemas/PowerMetric'
          title: metrics
          type: array
      required:
      - metrics
      - user_id
      title: PowerMetrics
      type: object
    InstantPowerMetric:
      example:
        user_id: userId1
        units: kW
        value: 12.5
        timestamp: 2023-01-20T00:00:00Z
      properties:
        user_id:
          example: userId1
          maxLength: 10
          minLength: 10
          title: user_id
          type: string
        timestamp:
          example: 2023-01-20T00:00:00Z
          format: date-time
          title: timestamp
          type: string
        value:
          example: 12.5
          format: float
          minimum: 0
          title: value
          type: number
        units:
          $ref: '#/components/schemas/PowerUnits'
      required:
      - timestamp
      - units
      - user_id
      - value
      title: InstantPowerMetric
      type: object
...
    Error:
      description: Error model
      example:
        error: error
      properties:
        error:
          title: error
          type: string
      required:
      - error
      title: Error
      type: object
  securitySchemes:
    Bearer:
      bearerFormat: JWT
      scheme: bearer
      type: http
      x-bearerInfoFunc: statistics_manager_service.controllers.security_controller.info_from_Bearer